ljx project/requirements.txt
tensorflow
numpy
matplotlib
pandas
scikit-learn
Flask
Pillow  # If you're using PIL for image processing